split class to new file and hope it works with the callback(monitor)
